QR Code
QR Code Generator

This repository offers a simple Python script for creating a QR code with custom colors. The script use the qrcode library and is intended to function seamlessly within a Jupyter Notebook environment. It creates a QR code from the data provided and shows the QR code image right in the notebook.

Features
QR codes can be generated with custom fill and background colors.
Displays the generated QR code picture in a Jupyter notebook.

Usage
Clone this repository onto your local workstation.
Open the Jupyter Notebook where you intend to run the script.
Copy and paste the code below into a notepad cell.Saves the QR code image as a file.